cli: change inat_taxon_for_taxon() to return an id
Don't return the whole inat taxon struct as we only need the ID.

+++ src/inaturalist.rs
@@ -18,7 +18,7 @@ pub enum Error {
 
 #[derive(Deserialize, Debug)]
 pub struct InaturalistTaxon {
-    pub id: u32,
+    pub id: u64,
     pub name: String,
     pub rank: String,
     pub is_active: bool,
@@ -90,7 +90,7 @@ pub async fn find_taxon(
 
 pub async fn fetch_seed_observations(
     client: &reqwest::Client,
-    taxon_id: u32,
+    taxon_id: u64,
     location: &SearchArea,
 ) -> Result<Vec<ObservationDate>, Error> {
     let mut observations: Vec<ObservationDate> = Vec::new();
